+++ mod_ci/forms.py
@@ -4,13 +4,11 @@ from wtforms.validators import DataRequired
 
 
 class AddUsersToBlacklist(Form):
-    userID = IntegerField('User ID', [DataRequired(
-        message='GitHub User ID not filled in')])
+    userID = IntegerField('User ID', [DataRequired(message='GitHub User ID not filled in')])
     comment = StringField('Comment')
     submit = SubmitField('Add User')
 
 
 class RemoveUsersFromBlacklist(Form):
-    userID = IntegerField('User ID', [DataRequired(
-        message='GitHub User ID not filled in')])
+    userID = IntegerField('User ID', [DataRequired(message='GitHub User ID not filled in')])
     submit = SubmitField('Remove User')
