+++ templates/ci/blocked_users.html
@@ -32,7 +32,7 @@
     <div class="row">
         <h1>Add Users</h1>
     </div>
-    <form method="post" name="registrationForm" id="registrationForm" action="{{ url_for('.add_user_to_blacklist') }}">
+    <form method="post" name="addUserForm" id="addUserForm" action="{{ url_for('.add_user_to_blacklist') }}">
         {{ form.csrf_token }}
         {% with messages = get_flashed_messages(with_categories=true) %}
             {% if messages %}
@@ -68,10 +68,12 @@
             <div class="medium-12 columns">
                 {{ macros.render_field(form.submit) }}
             </div>
+        </div>
+    </form>
     <div class="row">
         <h1>Remove Blocked Users</h1>
     </div>
-    <form method="post" name="registrationForm" id="registrationForm" action="{{ url_for('.remove_user_from_blacklist') }}">
+    <form method="post" name="removeUserForm" id="removeUserForm" action="{{ url_for('.remove_user_from_blacklist') }}">
         {{ form.csrf_token }}
         {% with messages = get_flashed_messages(with_categories=true) %}
             {% if messages %}
@@ -105,4 +107,5 @@
                 {{ macros.render_field(form.submit) }}
             </div>
         </div>
+    </form>
 {% endblock %}
